WIP #1887: Add AbstractSyntheticRecordPlannerTest base class for tests
